Добавлена возможность сделать обратную матрицу. Тип данных в ячейках матрицы теперь double. Исправлена ошибка с округлением

+++ RSPO_UP_5/Model/MatrixConverter.cs
@@ -30,7 +30,7 @@ namespace RSPO_UP_5.Model
             {
                 for (int j = 0; j < table.Columns.Count; j++)
                 {
-                    matrix[i, j] = int.Parse(table.Rows[i][j].ToString());
+                    matrix[i, j] = double.Parse(table.Rows[i][j].ToString());
                 }
             }
